% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eidcodes}
\alias{eidcodes}
\title{EID Codes}
\format{
\tabular{ll}{
\bold{Variable} \tab \bold{Description} \cr
eid    \tab EID Code \cr
city   \tab City Name\cr
county \tab County
}
}
\source{
\url{https://www.osa.state.mn.us/reports-data-analysis/data/city-comparison-tool/}
}
\usage{
eidcodes
}
\description{
List of EID Codes used by the Minnesota Office of the State Auditor
}
\examples{
data(eidcodes)
}
\keyword{datasets}
