% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osadata.R
\name{get_osadata}
\alias{get_osadata}
\title{Obtain OSA Data (webscrape)}
\usage{
get_osadata(eid, years = 2012:2021)
}
\arguments{
\item{eid}{EID Code (see \code{eidcodes})}

\item{years}{vector of years (2012 to 2021 available)}
}
\value{
tibble of requested financial data
}
\description{
Obtain OSA Data (webscrape)
}
