% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnosa-package.R
\docType{package}
\name{mnosa-package}
\alias{mnosa}
\alias{mnosa-package}
\title{mnosa: Minnesota Office of the State Auditor Data and Tools}
\description{
Data package and tools for interacting with city financial information from the MN Office of the State Auditor.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
