% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OSADebt}
\alias{OSADebt}
\title{Government Debt}
\format{
\tabular{ll}{
\bold{Variable} \tab \bold{Description} \cr
filename                \tab Original file name the data is from                     \cr
entity_id               \tab Unique identifier for city. This is the same as the eid.\cr
entity_name             \tab City name                                               \cr
parent_entity_name      \tab County name                                             \cr
financial_year          \tab Year of data                                            \cr
class_code              \tab City classification code based on census population     \cr
population              \tab Population                                              \cr
gen_obligation          \tab General obligation debt                                 \cr
gen_ob_tax_incr         \tab General obligation tax increment financing              \cr
revenue_tax_incr        \tab Revenue from tax increment financing                    \cr
spec_assessment         \tab Special assessment revenue                              \cr
gen_ob_revenue          \tab General obligation revenue                              \cr
revenue                 \tab Total revenue                                           \cr
other                   \tab Other revenue                                           \cr
bond_os_begin_yr        \tab bond_os_begin_yr                                        \cr
bond_adjustment         \tab bond_adjustment                                         \cr
bond_issue_yr           \tab bond_issue_yr                                           \cr
bond_paid_yr            \tab bond_paid_yr                                            \cr
refunding_bonds         \tab refunding_bonds                                         \cr
olt_begin               \tab olt_begin                                               \cr
olt_adjustment          \tab olt_adjustment                                          \cr
olt_issued              \tab olt_issued                                              \cr
olt_paid                \tab olt_paid                                                \cr
short_term_begin        \tab short_term_begin                                        \cr
short_term_issued       \tab short_term_issued                                       \cr
short_term_paid         \tab short_term_paid                                         \cr
workers_comp            \tab workers_comp                                            \cr
comp_absence            \tab comp_absence                                            \cr
gen_fund_res            \tab gen_fund_res                                            \cr
gen_fund_desig          \tab gen_fund_desig                                          \cr
gen_fund_undesig        \tab gen_fund_undesig                                        \cr
spec_rev_res            \tab spec_rev_res                                            \cr
spec_rev_desig          \tab spec_rev_desig                                          \cr
spec_rev_undesig        \tab spec_rev_undesig                                        \cr
debt_svc_res            \tab debt_svc_res                                            \cr
debt_svc_desig          \tab debt_svc_desig                                          \cr
debt_svc_undesig        \tab debt_svc_undesig                                        \cr
cap_proj_res            \tab cap_proj_res                                            \cr
cap_proj_undesig        \tab cap_proj_undesig                                        \cr
cap_proj_desig          \tab cap_proj_desig                                          \cr
ta_reserve              \tab ta_reserve                                              \cr
ta_desig                \tab ta_desig                                                \cr
ta_undesig              \tab ta_undesig                                              \cr
tcu_reserve             \tab tcu_reserve                                             \cr
tcu_desig               \tab tcu_desig                                               \cr
tcu_undesig             \tab tcu_undesig                                             \cr
entr_fnd_cap_asset      \tab entr_fnd_cap_asset                                      \cr
entr_fnd_reserv         \tab entr_fnd_reserv                                         \cr
entr_fnd_unreserv       \tab entr_fnd_unreserv                                       \cr
gaap_ind                \tab gaap_ind                                                \cr
cash_balance            \tab cash_balance                                            \cr
other_investmt          \tab other_investmt                                          \cr
fed_securities          \tab fed_securities                                          \cr
interfund_beg_yr        \tab interfund_beg_yr                                        \cr
interfund_issue         \tab interfund_issue                                         \cr
interfund_paid          \tab interfund_paid                                          \cr
expr1028                \tab expr1028                                                \cr
expr1029                \tab expr1029                                                \cr
expr1030                \tab expr1030                                                \cr
gen_fund_cash           \tab gen_fund_cash                                           \cr
gen_fund_fed_sec        \tab gen_fund_fed_sec                                        \cr
gen_fund_oth_inv        \tab gen_fund_oth_inv                                        \cr
spec_rev_cash           \tab spec_rev_cash                                           \cr
spec_rev_fed_sec        \tab spec_rev_fed_sec                                        \cr
spec_rev_oth_inv        \tab spec_rev_oth_inv                                        \cr
opeb_liability          \tab opeb_liability                                          \cr
intl_svc_cap_asset      \tab intl_svc_cap_asset                                      \cr
intl_svc_unresrv        \tab intl_svc_unresrv                                        \cr
intl_svc_reserv         \tab intl_svc_reserv                                         \cr
priv_purp_tr_cap_asset  \tab priv_purp_tr_cap_asset                                  \cr
priv_purp_tr_reserve    \tab priv_purp_tr_reserve                                    \cr
priv_purp_tr_un_reserve \tab priv_purp_tr_un_reserve                                 \cr
tcure_cap_asset         \tab tcure_cap_asset                                         \cr
tcure_reserv            \tab tcure_reserv                                            \cr
tcure_unreserv          \tab tcure_unreserv
}
}
\source{
\url{https://www.osa.state.mn.us/reports-data-analysis/local-government/cities/}
}
\usage{
OSADebt
}
\description{
Historical data (2008-2021) of MN city government debt from the Office of the State Auditor.
}
\examples{
data(OSADebt)
}
\keyword{datasets}
