% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OSAEmpData}
\alias{OSAEmpData}
\title{Employee Data}
\format{
\tabular{ll}{
\bold{Variable} \tab \bold{Description} \cr
filename              \tab filename             \cr
entity_id             \tab entity_id            \cr
entity_name           \tab entity_name          \cr
class_code            \tab class_code           \cr
financial_year        \tab financial_year       \cr
population            \tab population           \cr
gen_govt_ft_ct        \tab gen_govt_ft_ct       \cr
gen_govt_pt_ct        \tab gen_govt_pt_ct       \cr
gen_govt_saly_amt     \tab gen_govt_saly_amt    \cr
gen_govt_benf_amt     \tab gen_govt_benf_amt    \cr
police_ft_ct          \tab police_ft_ct         \cr
police_pt_ct          \tab police_pt_ct         \cr
police_salary_amt     \tab police_salary_amt    \cr
police_benf_amt       \tab police_benf_amt      \cr
fire_ft_ct            \tab fire_ft_ct           \cr
fire_pt_ct            \tab fire_pt_ct           \cr
fire_salaries_amt     \tab fire_salaries_amt    \cr
fire_benf_amt         \tab fire_benf_amt        \cr
streets_ft_ct         \tab streets_ft_ct        \cr
streets_pt_ct         \tab streets_pt_ct        \cr
streets_sal_amt       \tab streets_sal_amt      \cr
streets_benf_amt      \tab streets_benf_amt     \cr
sanitation_ft_ct      \tab sanitation_ft_ct     \cr
sanitation_pt_ct      \tab sanitation_pt_ct     \cr
sanitation_saly_amt   \tab sanitation_saly_amt  \cr
sanitation_benf_amt   \tab sanitation_benf_amt  \cr
library_ft_ct         \tab library_ft_ct        \cr
library_pt_ct         \tab library_pt_ct        \cr
library_saly_amt      \tab library_saly_amt     \cr
library_benf_amt      \tab library_benf_amt     \cr
park_and_rec_ft_ct    \tab park_and_rec_ft_ct   \cr
park_and_rec_pt_ct    \tab park_and_rec_pt_ct   \cr
park_and_rec_saly_amt \tab park_and_rec_saly_amt\cr
park_and_rec_benf_amt \tab park_and_rec_benf_amt\cr
all_other_ft_ct       \tab all_other_ft_ct      \cr
all_other_pt_ct       \tab all_other_pt_ct      \cr
all_other_saly_amt    \tab all_other_saly_amt   \cr
all_other_benf_amt    \tab all_other_benf_amt   \cr
health_ft_ct          \tab health_ft_ct         \cr
health_pt_ct          \tab health_pt_ct         \cr
health_saly_amt       \tab health_saly_amt      \cr
health_ben_amt        \tab health_ben_amt       \cr
conserv_ft_ct         \tab conserv_ft_ct        \cr
conserv_pt_ct         \tab conserv_pt_ct        \cr
conserv_saly_amt      \tab conserv_saly_amt     \cr
conserv_ben_amt       \tab conserv_ben_amt      \cr
econ_dev_ft_ct        \tab econ_dev_ft_ct       \cr
econ_dev_pt_ct        \tab econ_dev_pt_ct       \cr
econ_dev_saly_amt     \tab econ_dev_saly_amt    \cr
econ_dev_ben_amt      \tab econ_dev_ben_amt     \cr
enterprise_ft_ct      \tab enterprise_ft_ct     \cr
enterprise_pt_ct      \tab enterprise_pt_ct     \cr
enterprise_saly_amt   \tab enterprise_saly_amt  \cr
enterprise_benf_amt   \tab enterprise_benf_amt  \cr
num_of_retirees       \tab num_of_retirees      \cr
retire_benefits       \tab retire_benefits
}
}
\source{
\url{https://www.osa.state.mn.us/reports-data-analysis/local-government/cities/}
}
\usage{
OSAEmpData
}
\description{
Historical data (2014-2021) of MN city employee data from the Office of the State Auditor.
}
\examples{
data(OSAEmpData)
}
\keyword{datasets}
