% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OSAEntFunds}
\alias{OSAEntFunds}
\title{Enterprise Funds}
\format{
\tabular{ll}{
\bold{Variable} \tab \bold{Description} \cr
filename                                \tab filename                               \cr
entity_id                               \tab entity_id                              \cr
entity_name                             \tab entity_name                            \cr
parent_entity_name                      \tab parent_entity_name                     \cr
name_of_enterprise                      \tab name_of_enterprise                     \cr
enterprise_category                     \tab enterprise_category                    \cr
enterprise_category_id                  \tab enterprise_category_id                 \cr
foot_notes                              \tab foot_notes                             \cr
financial_year                          \tab financial_year                         \cr
class_code                              \tab class_code                             \cr
population                              \tab population                             \cr
operating_revenue                       \tab operating_revenue                      \cr
cost_of_sales                           \tab cost_of_sales                          \cr
gross_profit                            \tab gross_profit                           \cr
operating_expenses                      \tab operating_expenses                     \cr
operating_income                        \tab operating_income                       \cr
nonoperating_revenue                    \tab nonoperating_revenue                   \cr
nonoperating_expenses                   \tab nonoperating_expenses                  \cr
net_income                              \tab net_income                             \cr
taxes                                   \tab taxes                                  \cr
federal_grants                          \tab federal_grants                         \cr
state_grants                            \tab state_grants                           \cr
count_local_grants                      \tab count_local_grants                     \cr
transfer_in                             \tab transfer_in                            \cr
transfer_out                            \tab transfer_out                           \cr
total_cap_outlay                        \tab total_cap_outlay                       \cr
borrowing                               \tab borrowing                              \cr
interest_paid                           \tab interest_paid                          \cr
bonds_paid                              \tab bonds_paid                             \cr
principal_payments_other_long_term_debt \tab principal_payments_other_long_term_debt\cr
bonds_os_yr_end                         \tab bonds_os_yr_end                        \cr
oltos_yr_end                            \tab oltos_yr_end                           \cr
depreciation                            \tab depreciation                           \cr
int_fnd_debt_borrow                     \tab int_fnd_debt_borrow                    \cr
int_fnd_debt_paid                       \tab int_fnd_debt_paid                      \cr
city_class                              \tab city_class                             \cr
operating_exp                           \tab operating_exp                          \cr
operating_income_loss                   \tab operating_income_loss                  \cr
completed_date                          \tab completed_date
}
}
\source{
\url{https://www.osa.state.mn.us/reports-data-analysis/local-government/cities/}
}
\usage{
OSAEntFunds
}
\description{
Historical data (2008-2021) of MN city enterprise funds from the Office of the State Auditor.
}
\examples{
data(OSAEntFunds)
}
\keyword{datasets}
