% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OSAFundBal}
\alias{OSAFundBal}
\title{Fund Balance}
\format{
\tabular{ll}{
\bold{Variable} \tab \bold{Description} \cr
filename                      \tab filename                     \cr
entity_id                     \tab entity_id                    \cr
entity_name                   \tab entity_name                  \cr
parent_entity_name            \tab parent_entity_name           \cr
financial_year                \tab financial_year               \cr
class_code                    \tab class_code                   \cr
general_restricted            \tab general_restricted           \cr
general_committed             \tab general_committed            \cr
general_assigned              \tab general_assigned             \cr
general_unassigned            \tab general_unassigned           \cr
special_revenue_restricted    \tab special_revenue_restricted   \cr
special_revenue_committed     \tab special_revenue_committed    \cr
special_revenue_assigned      \tab special_revenue_assigned     \cr
special_revenue_unassigned    \tab special_revenue_unassigned   \cr
debt_service_restricted       \tab debt_service_restricted      \cr
debt_service_committed        \tab debt_service_committed       \cr
debt_service_assigned         \tab debt_service_assigned        \cr
debt_service_unassigned       \tab debt_service_unassigned      \cr
capital_projects_restricted   \tab capital_projects_restricted  \cr
capital_projects_committed    \tab capital_projects_committed   \cr
capital_projects_assigned     \tab capital_projects_assigned    \cr
capital_projects_unassigned   \tab capital_projects_unassigned  \cr
permanent_restricted          \tab permanent_restricted         \cr
permanent_committed           \tab permanent_committed          \cr
permanent_assigned            \tab permanent_assigned           \cr
permanent_unassigned          \tab permanent_unassigned         \cr
component_units_restricted    \tab component_units_restricted   \cr
component_units_committed     \tab component_units_committed    \cr
component_units_assigned      \tab component_units_assigned     \cr
component_units_unassigned    \tab component_units_unassigned   \cr
general_nonspendable          \tab general_nonspendable         \cr
special_revenue_nonspendable  \tab special_revenue_nonspendable \cr
debt_service_nonspendable     \tab debt_service_nonspendable    \cr
capital_projects_nonspendable \tab capital_projects_nonspendable\cr
permanent_nonspendable        \tab permanent_nonspendable       \cr
component_units_nonspendable  \tab component_units_nonspendable \cr
gaap_ind                      \tab gaap_ind
}
}
\source{
\url{https://www.osa.state.mn.us/reports-data-analysis/local-government/cities/}
}
\usage{
OSAFundBal
}
\description{
Historical data (2011-2021) of MN city fund balances from the Office of the State Auditor.
}
\examples{
data(OSAFundBal)
}
\keyword{datasets}
