% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OSAGovFunds}
\alias{OSAGovFunds}
\title{Government Funds}
\format{
\tabular{ll}{
\bold{Variable} \tab \bold{Description} \cr
filename                                         \tab Original file name the data is from                       \cr
entity_id                                        \tab Unique identifier for city. This is the same as the eid.  \cr
entity_name                                      \tab City name                                                 \cr
parent_entity_name                               \tab County name                                               \cr
entity_type                                      \tab Descriptor of type of entity                              \cr
class_code                                       \tab City classification code based on census population       \cr
population                                       \tab Estimated population                                      \cr
financial_year                                   \tab Year of data                                              \cr
tax_capacity                                     \tab Tax capacity                                              \cr
tax_levy                                         \tab Total tax levy                                            \cr
special_assessment_levy                          \tab Special assessment levy                                   \cr
property_taxes                                   \tab Property taxes collected for general government           \cr
tax_increments                                   \tab TIF monies collected                                      \cr
franchise_fees                                   \tab Franchise fees collected                                  \cr
sales_tax                                        \tab Sales taxes collected (for city)                          \cr
hotel_motel_tax                                  \tab Hotel and motel taxes collected                           \cr
gambling_tax                                     \tab Gambling taxes                                            \cr
gravel_tax                                       \tab Gravel taxes                                              \cr
special_assessments                              \tab Special Assessments collected                             \cr
licenses_and_permits                             \tab Licenses and permit fees collected                        \cr
federal_cdbg_grants                              \tab Federal Community Development Block Grants                \cr
federal_education_grants                         \tab Federal Education Grants                                  \cr
federal_transportation_grants                    \tab Federal Transportation Grants                             \cr
federal_human_services_grants                    \tab Federal Human Services Grants                             \cr
federal_emergency_management_aid                 \tab Federal Emergency Management Aid                          \cr
all_other_federal_grants                         \tab All Other Federal Grants                                  \cr
total_federal_grants                             \tab Total Federal Grants                                      \cr
state_local_government_aid                       \tab Local Government Aid (LGA)                                \cr
state_market_value_credit_real                   \tab Market Value Credit                                       \cr
state_market_value_credit_ag                     \tab Market Value Credit for Agriculture                       \cr
state_taconite_homestead_credit                  \tab Taconite Homestead Credit                                 \cr
state_thc                                        \tab State THC funds                                           \cr
state_taconite_aids                              \tab State Taconite Aid                                        \cr
state_county_program_aid                         \tab State County Program Aid                                  \cr
state_manufactured_home_homestead_credit         \tab Manufactured Home Homestead Credit                        \cr
state_attached_machinery_aid                     \tab Attached Machinery Aid                                    \cr
state_disparity_reduction_aid                    \tab Disparity Reduction Aid                                   \cr
state_transportation_grants                      \tab Transportation Grants                                     \cr
state_human_service_grants                       \tab Human Service Grants                                      \cr
state_criminal_justice_aid                       \tab Criminal Justice Aid                                      \cr
state_pera_aid                                   \tab Public Employee Retirement Account Aid                    \cr
state_highway_grants                             \tab State Highway Grants                                      \cr
state_education_grants                           \tab State Education Grants                                    \cr
state_police_aid                                 \tab State Police Aid                                          \cr
state_lpa                                        \tab State LPA                                                 \cr
state_haca                                       \tab State HACA                                                \cr
all_other_state_grants                           \tab All Other State Grants                                    \cr
total_state_grants                               \tab Total State Grants                                        \cr
county_highway_grants                            \tab County Highway Grants                                     \cr
all_other_county_grants                          \tab All Other County Grants                                   \cr
local_irrrb_grants                               \tab Local Iron Range Resources and Rehabilitation Board Grants\cr
all_other_local_grants                           \tab All Other Local Grants                                    \cr
total_intergovernmental_revenues                 \tab Total Intergovernmental Revenues                          \cr
general_government_fees                          \tab General Government Fees                                   \cr
police_and_fire_contracts                        \tab Police and Fire Contracts Revenue                         \cr
other_public_safety_fees                         \tab Other Public Safety Fees                                  \cr
street_and_highway_fees                          \tab Street and Highway Fees                                   \cr
sanitation_fees                                  \tab Sanitation Fees                                           \cr
library_fees                                     \tab Library Fees                                              \cr
park_and_recreation_fees                         \tab Park and Recreation Fees                                  \cr
airport_fees                                     \tab Airport Fees                                              \cr
transit_fees                                     \tab Transit Fees                                              \cr
cemetery_fees                                    \tab Cemetery Fees                                             \cr
all_other_service_charges                        \tab All other service charges                                 \cr
total_charges_for_services                       \tab Total charges for services                                \cr
fines_and_forfeits                               \tab Fines and Forfeitures                                     \cr
interest_earnings                                \tab Interest earnings                                         \cr
all_other_revenue                                \tab All Other Revenue                                         \cr
total_revenues                                   \tab Total revenues                                            \cr
bonds_issued                                     \tab Bonds issues                                              \cr
other_long_term_debt_issued                      \tab Other long term debt issued                               \cr
short_term_debt_issued                           \tab Short term debt issued                                    \cr
other_financing_sources                          \tab Other financing sources                                   \cr
transfers_from_enterprise_funds                  \tab Transfers from enterprise funds                           \cr
transfers_from_governmental_funds                \tab Transfers from governmental funds                         \cr
total_revenues_other_sources                     \tab Total revenues from other sources                         \cr
governing_board_current_expend                   \tab Governening board expenditures                            \cr
administration_and_finance_current_expend        \tab Expenditures for administration and finance               \cr
all_other_general_government_current_expend      \tab All other general government current expenditures         \cr
general_government_capital_outlay                \tab General government capital outlay                         \cr
police_sheriff_current_expend                    \tab Police and sheriff current expenditures                   \cr
police_sheriff_capital_outlay                    \tab Police and sheriff capital outlay                         \cr
corrections_current_expend                       \tab Corrections and jail current expenditure                  \cr
corrections_capital_outlay                       \tab Corrections and jail capital outlay                       \cr
ambulance_current_expend                         \tab Ambulance current expenditures                            \cr
ambulance_capital_outlay                         \tab Ambulance capital outlay                                  \cr
fire_current_expend                              \tab Fire current expenditures                                 \cr
fire_capital_outlay                              \tab Fire capital outlay                                       \cr
all_other_public_safety_current_expend           \tab All other public safety current expenditures              \cr
all_other_public_safety_capital_outlay           \tab All other public safety capital outlay                    \cr
total_public_safety_capital_outlay               \tab Total public safety capital outlay                        \cr
street_highway_maintenance                       \tab Street highway maintenance                                \cr
snow_and_ice_removal                             \tab Snow and ice removal                                      \cr
street_highway_engineering                       \tab Street highway engineering                                \cr
street_lighting                                  \tab Street lighting                                           \cr
street_highway_construction                      \tab Street highway construction                               \cr
all_other_street_highway_capital_outlay          \tab All other street highway capital outlay                   \cr
refuse_collection_and_disposal_current_expend    \tab Refuse collection and disposal current expenditures       \cr
all_other_sanitation_current_expend              \tab All other sanitation current expenditures                 \cr
sanitation_capital_outlay                        \tab Sanitation capital outlay                                 \cr
health_current_expend                            \tab Health current expenditures                               \cr
health_capital_outlay                            \tab Heatlh capital outlay                                     \cr
library_current_expenditures                     \tab Library current expenditures                              \cr
library_capital_outlay                           \tab Library capital outlay                                    \cr
park_recreation_current_expend                   \tab Park and recreation current expenditure                   \cr
park_recreation_capital_outlay                   \tab Park and recreation capital outlay                        \cr
housing_urban_redevelopment_current_expend       \tab Housing and urban redevelopment current expenditures      \cr
housing_urban_redevelopment_capital_outlay       \tab Housing and urban redevelopment capital outlay            \cr
economic_development_current_expenditures        \tab Economic development current expenditures                 \cr
economic_development_capital_outlay              \tab Economic development capital outlay                       \cr
conservation_of_natural_resources_current_expend \tab Conservation of natural resources current expenditures    \cr
conservation_of_natural_resources_capital_outlay \tab Conservation of natural resources capital outlay          \cr
airport_current_expend                           \tab Airport current expenditures                              \cr
airport_capital_outlay                           \tab Airport capital outlay                                    \cr
transit_current_expend                           \tab Transit current expenditures                              \cr
transit_capital_outlay                           \tab Transit capital outlay                                    \cr
cemetery_current_expend                          \tab Cemetary current expenditures                             \cr
cemetery_capital_outlay                          \tab Cemetary capital outlay                                   \cr
unallocated_pension_costs                        \tab Unallocated pension costs                                 \cr
unallocated_insurance_costs                      \tab Unallocated insurance costs                               \cr
all_other_current_expend                         \tab All other current expenditures                            \cr
all_other_capital_outlay                         \tab All other capital outlay                                  \cr
capital_outlay_for_enterprise_funds              \tab Capital outlay for enterprise funds                       \cr
total_current_expenditures                       \tab Total current expenditures                                \cr
total_capital_outlay                             \tab Total capital outlay                                      \cr
bond_principal_payments                          \tab Bond principal payments                                   \cr
other_long_term_debt_principal_payments          \tab Other long term debt principal payments                   \cr
interest_payments_fiscal_charges                 \tab Interest payments fiscal charges                          \cr
total_expenditures                               \tab Total expenditures                                        \cr
principal_payments_refunded_bonds                \tab Principal payments on refunded bonds                      \cr
other_financing_uses                             \tab Other financing uses                                      \cr
transfers_to_enterprise_funds                    \tab Transfers to enterprise funds                             \cr
transfers_to_governmental_funds                  \tab Transfers to governmental funds                           \cr
total_expenditures_other_uses                    \tab Total expenditures other uses                             \cr
tn_water_ce                                      \tab Town water capital expense                                \cr
tn_water_co                                      \tab Town water capital outlay                                 \cr
tn_water_charge                                  \tab Town water charges                                        \cr
admin_fines                                      \tab Admin fines                                               \cr
street_highway_administration                    \tab Street highway administration                             \cr
human_service_income_maintenance                 \tab Human service income maintenance                          \cr
human_services_social_services                   \tab Human services social services                            \cr
all_other_human_services_current_expenditures    \tab All other human services current expenditures             \cr
human_services_capital_outlay                    \tab Human services capital outlay                             \cr
education_current_expend                         \tab Education current expenditures                            \cr
education_capital_outlay                         \tab Education capital outlay                                  \cr
alt_current_expend                               \tab Alternative current expenditures                          \cr
gaap_ind                                         \tab Generally Accepted Accounting Practices indicator         \cr
fed_coronavirus_relief_funds                     \tab Federal Coronavirus Relief Funds
}
}
\source{
\url{https://www.osa.state.mn.us/reports-data-analysis/local-government/cities/}
}
\usage{
OSAGovFunds
}
\description{
Historical data (2008-2021) of MN city government funds from the Office of the State Auditor.
}
\examples{
data(OSAGovFunds)
}
\keyword{datasets}
